+++ apps/mobile/src/app/components/ShareLocationConfirmModal/ShareLocation/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ShareLocation';
