+++ apps/mobile/src/app/components/ShareLocationConfirmModal/ShareLocationConfirmModal.tsx
@@ -1,59 +1,25 @@
-import { useChatSending } from '@mezon/core';
-import { Icons } from '@mezon/mobile-components';
-import { useTheme } from '@mezon/mobile-ui';
-import { selectChannelById, selectDmGroupCurrent, selectGeolocation } from '@mezon/store-mobile';
-import { IMessageSendPayload, filterEmptyArrays, processText } from '@mezon/utils';
+import { referencesActions, selectGeolocation } from '@mezon/store-mobile';
 import { ChannelStreamMode } from 'mezon-js';
-import React, { useEffect, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Text, TouchableOpacity, View } from 'react-native';
+import React, { useCallback, useEffect, useState } from 'react';
 import Modal from 'react-native-modal';
-import { useSelector } from 'react-redux';
-import { style } from './styles';
+import { useDispatch, useSelector } from 'react-redux';
+import ShareLocation from './ShareLocation';
 
 const ShareLocationConfirmModal = ({ mode, channelId }: { mode: ChannelStreamMode; channelId: string }) => {
-	const { themeValue } = useTheme();
-	const styles = style(themeValue);
 	const geoLocation = useSelector(selectGeolocation);
 	const [visible, setVisible] = useState<boolean>(false);
-	const currentChannel = useSelector(selectChannelById(channelId));
-	const currentDmGroup = useSelector(selectDmGroupCurrent(channelId));
-	const [links, setLinks] = useState([]);
-	const { t } = useTranslation('message');
+	const dispatch = useDispatch();
 
-	const [googleMapsLink, setGoogleMapsLink] = useState<string>('');
-	const { sendMessage } = useChatSending({
-		mode,
-		channelOrDirect:
-			mode === ChannelStreamMode.STREAM_MODE_CHANNEL || mode === ChannelStreamMode.STREAM_MODE_THREAD ? currentChannel : currentDmGroup
-	});
 	useEffect(() => {
 		if (geoLocation) {
-			const { latitude, longitude } = geoLocation;
-			const googleMapsLink = `https://www.google.com/maps?q=${latitude},${longitude}&z=14&t=m&mapclient=embed`;
-			setGoogleMapsLink(googleMapsLink);
-			const { links } = processText(googleMapsLink);
-			setLinks(links);
 			setVisible(true);
 		}
 	}, [geoLocation]);
 
-	const handleSendMessage = async () => {
-		const payloadSendMessage: IMessageSendPayload = {
-			t: googleMapsLink,
-			hg: [],
-			ej: [],
-			lk: links || [],
-			mk: [],
-			vk: []
-		};
+	const handelCancelModal = useCallback(() => {
 		setVisible(false);
-		await sendMessage(filterEmptyArrays(payloadSendMessage), [], [], [], false, false, true);
-	};
-
-	const handelCancelModal = () => {
-		setVisible(false);
-	};
+		dispatch(referencesActions.setGeolocation(null));
+	}, []);
 
 	return (
 		<Modal
@@ -65,32 +31,7 @@ const ShareLocationConfirmModal = ({ mode, channelId }: { mode: ChannelStreamMod
 			avoidKeyboard={false}
 			backdropColor={'rgba(0,0,0, 0.7)'}
 		>
-			<View style={styles.modalContainer}>
-				<View style={styles.modalHeader}>
-					<Text style={styles.headerText}>
-						{t('shareLocationModal.sendThisLocation')}{' '}
-						{mode === ChannelStreamMode.STREAM_MODE_CHANNEL || mode === ChannelStreamMode.STREAM_MODE_THREAD
-							? currentChannel?.channel_label
-							: currentDmGroup?.channel_label}
-					</Text>
-				</View>
-				<View style={styles.modalContent}>
-					<View style={styles.circleIcon}>
-						<Icons.LocationIcon />
-					</View>
-					<Text
-						style={styles.textContent}
-					>{`${t('shareLocationModal.coordinate')} (${geoLocation?.latitude}, ${geoLocation?.longitude})`}</Text>
-				</View>
-				<View style={styles.modalFooter}>
-					<TouchableOpacity style={styles.button} onPress={handelCancelModal}>
-						<Text style={styles.textButton}>{t('shareLocationModal.cancel')}</Text>
-					</TouchableOpacity>
-					<TouchableOpacity style={styles.button} onPress={handleSendMessage}>
-						<Text style={styles.textButton}>{t('shareLocationModal.send')}</Text>
-					</TouchableOpacity>
-				</View>
-			</View>
+			<ShareLocation oncancel={handelCancelModal} mode={mode} channelId={channelId} geoLocation={geoLocation} />
 		</Modal>
 	);
 };
